+++ R/owid_get_catalog.R
@@ -32,22 +32,12 @@ owid_get_catalog <- function(
         read.csv() |>
         tibble::as_tibble()
 
-      # Parse logicals
-      catalog_raw$isInheritanceEnabled <- catalog_raw$isInheritanceEnabled ==
-        "True"
-      catalog_raw$isIndexable <- catalog_raw$isIndexable == "True"
-      catalog_raw$isPublished <- catalog_raw$isPublished == "True"
-
-      # Parse dates
-      catalog_raw$createdAt <- as.Date(catalog_raw$createdAt)
-      catalog_raw$updatedAt <- as.Date(catalog_raw$updatedAt)
-      catalog_raw$lastEditedAt <- as.Date(catalog_raw$lastEditedAt)
-      catalog_raw$publishedAt <- as.Date(catalog_raw$publishedAt)
+      catalog_parsed <- parse_catalog_columns(catalog_raw)
 
       if (snake_case) {
-        catalog <- to_snake_case(catalog_raw)
+        catalog <- to_snake_case(catalog_parsed)
       } else {
-        catalog <- catalog_raw
+        catalog <- catalog_parsed
       }
 
       catalog